+++ SCADA_Core/DigitalInput.cs
@@ -15,7 +15,7 @@ namespace SCADA_Core
         }
 
         public DigitalInput(string tagName, string description, string driver, int ioAddress,
-                            float scanTime, bool enableScan, bool manualMode)
+                            int scanTime, bool enableScan, bool manualMode)
             : base(tagName, description, driver, ioAddress, scanTime, enableScan, manualMode)
         {
         }
